tools/opai-marq/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
httpx>=0.25.0
python-jose[cryptography]>=3.3.0
cryptography>=41.0.0
croniter>=1.4.0
PyJWT>=2.8.0
google-auth>=2.23.0
Pillow>=10.0.0
python-multipart>=0.0.9
python-slugify>=8.0.0
